AquaSense/requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
httpx==0.23.0
python-dotenv==0.19.0
supabase==1.0.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.5
torch==2.0.0
torchvision==0.15.0
pillow==9.5.0
opencv-python==4.7.0.72
numpy==1.24.3
ultralytics==8.0.0
aoifiles==0.0.1
groq==0.1.0
pydantic==1.10.2
googleapi==0.1.0
FAISS==1.16.0
scikit-learn==1.2.2
sentence-transformers==2.2.2
joblib==1.2.0